requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
httpx>=0.27.0
aiosqlite>=0.20.0
supabase>=2.4.0
Pillow>=10.3.0
yt-dlp>=2024.4.9
